LayoutTests/http/wpt/beacon/connect-src-beacon-blocked.sub-expected.txt

CONSOLE MESSAGE: Refused to connect to http://www.localhost:8800/common/text-plain.txt because it does not appear in the connect-src directive of the Content Security Policy.

PASS sendBeacon should not throw.
PASS redirect case
